# src/main/resources/public/swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "API para gerenciamento das pautas do Banco Sicredi",
		"version": "1.0",
		"title": "API de Votação",
		"contact": {
			
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "votacaosicredi.herokuapp.com",
	"basePath": "/",
	"tags": [
		{
			"name": "Pauta",
			"description": "API para gerencimento das pautas de votação do Banco Sicredi"
		}
	],
	"paths": {
		"/api/v1/pauta": {
			"post": {
				"tags": [
					"Pauta"
				],
				"summary": "Criar uma pauta",
				"operationId": "criarPautaUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"in": "body",
						"name": "Pauta",
						"description": "Parametros para criar uma Pauta",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CriarPauta"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Pauta criada"
					},
					"400": {
						"description": "Não foram passados os parâmetros corretos para criação da pauta"
					}
				},
				"deprecated": false
			},
			"get": {
				"tags": [
					"Pauta"
				],
				"summary": "Busca todas as pautas",
				"operationId": "buscarTodasAsPautasUsingGET",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Response«List«Pauta»»"
						}
					},
					"404": {
						"description": "Pauta não encontrada"
					}
				}
			}
		},
		"/api/v1/pauta/{id}": {
			"get": {
				"tags": [
					"Pauta"
				],
				"summary": "Busca uma pauta a partir do código",
				"operationId": "buscarPorIdUsingGET",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Código da pauta",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Response«Pauta»"
						}
					},
					"404": {
						"description": "Pauta não encontrada"
					}
				},
				"deprecated": false
			},
			"patch": {
				"tags": [
					"Pauta"
				],
				"summary": "Inicia a sessão de votação de uma pauta",
				"operationId": "iniciarSessaoUsingPATCH",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Código da pauta",
						"required": true,
						"type": "integer",
						"format": "int64"
					},
					{
						"name": "duracaoSessao",
						"in": "body",
						"description": "Duração da sessão a ser iniciada em minutos, caso não seja informado será considerado o valor 1 minuto",
						"required": false,
						"type": "integer",
						"format": "int64",
						"schema": {
							"$ref": "#/definitions/IniciarSessao"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Sessão aberta com sucesso"
					},
					"400": {
						"description": "Não foram passados os paramêtros corretos para abertura da sessãoo"
					}
				},
				"deprecated": false
			}
		},
		"/api/v1/pauta/{id}/votar": {
			"post": {
				"tags": [
					"Pauta"
				],
				"summary": "Vota em uma pauta",
				"operationId": "votarUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "Código da pauta",
						"required": true,
						"type": "integer",
						"format": "int64"
					},
					{
						"in": "body",
						"name": "Voto",
						"description": "Voto",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Voto"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Voto computado com sucesso"
					},
					"400": {
						"description": "Não foram passados os parâmentros corretos para computar o voto"
					}
				},
				"deprecated": false
			}
		}
	},
	"definitions": {
		"IniciarSessao": {
			"type": "object",
			"properties": {
				"duracaoSessao": {
					"type": "integer",
					"format": "int64"
				}
			},
			"title": "IniciarSessao"
		},
		"Response«List«Pauta»»": {
			"type": "object",
			"properties": {
				"data": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/Pauta"
					}
				},
				"errors": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"title": "Response«List«Pauta»»"
		},
		"Pauta": {
			"type": "object",
			"properties": {
				"assunto": {
					"type": "string"
				},
				"dataTerminoSessao": {
					"type": "string",
					"format": "date-time"
				},
				"duracaoSessao": {
					"type": "integer",
					"format": "int64"
				},
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"resultado": {
					"type": "string"
				},
				"sessaoEncerrada": {
					"type": "boolean"
				}
			},
			"title": "Pauta"
		},
		"CriarPauta": {
			"type": "object",
			"properties": {
				"assunto": {
					"type": "string"
				}
			},
			"title": "CriarPauta"
		},
		"Response«Pauta»": {
			"type": "object",
			"properties": {
				"data": {
					"$ref": "#/definitions/Pauta"
				},
				"errors": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"title": "Response«Pauta»"
		},
		"Response«Voto»": {
			"type": "object",
			"properties": {
				"data": {
					"$ref": "#/definitions/Voto"
				},
				"errors": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"title": "Response«Voto»"
		},
		"Voto": {
			"type": "object",
			"properties": {
				"idPauta": {
					"type": "integer",
					"format": "int64"
				},
				"idUsuario": {
					"type": "integer",
					"format": "int64"
				},
				"opcao": {
					"type": "string"
				}
			},
			"title": "Voto"
		}
	}
}